Use English for all names
Although this data is about Brazil and most of the people interested in it
would speak Portuguese, I'm also using it as an example on how to build a
datapackage and run data validations with goodtables, so it's useful to have it
in English.

I intend to keep the descriptions, and potentially comments in Portuguese.